requirements.txt
pillow
pygame
matplotlib
tensorflow